requirements.txt
wecom-aibot-sdk-python
aiohttp
